# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "커피 키오스크 API",
    "description": "키오스크·백오피스·마이페이지 API. 인증 필요 엔드포인트는 Authorize에 JWT를 입력하세요.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "로컬 개발" }
  ],
...
  "paths": {
...
    "/api/user/me": {
      "get": {
        "tags": ["User"],
        "summary": "내 정보 (포인트·마일리지·설정)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserMe" } } } },
          "401": { "description": "Unauthorized" },
          "404": { "description": "user not found" }
        }
      }
    },
    "/api/user/orders": {
      "get": {
        "tags": ["User"],
        "summary": "내 주문 내역",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "status", "in": "query", "schema": { "type": "string", "enum": ["ALL", "CANCELED", "WAITING", "COMPLETED", "PENDING"] }, "description": "ALL이 기본" }
        ],
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/UserOrderItem" } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/user/cart": {
      "get": {
        "tags": ["User"],
        "summary": "장바구니",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "성공", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array" } } } } } },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/user/update": {
      "patch": {
        "tags": ["User"],
        "summary": "이메일/비밀번호 변경 (비밀번호 변경 시 currentPassword 필수)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "currentPassword": { "type": "string" },
                  "newPassword": { "type": "string", "minLength": 6 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "success: true" },
          "400": { "description": "유효성 오류" },
          "401": { "description": "current password incorrect" },
          "409": { "description": "email already in use" }
        }
      }
    },
    "/api/user/settings": {
      "patch": {
        "tags": ["User"],
        "summary": "알림·마케팅 수신 설정",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "notificationEnabled": { "type": "boolean" },
                  "marketingConsent": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "성공" },
          "400": { "description": "provide notificationEnabled and/or marketingConsent" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "주문 생성 (키오스크)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["totalPrice", "items"],
                "properties": {
                  "totalPrice": { "type": "number" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": ["productId", "quantity"],
                      "properties": {
                        "productId": { "type": "string" },
                        "quantity": { "type": "number", "minimum": 1 },
                        "optionIds": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "주문 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderNumber": { "type": "string" },
                    "orderNo": { "type": "number" },
                    "orderId": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "유효성 오류 (INVALID_TOTAL_PRICE, ITEMS_REQUIRED 등)" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "로그인 응답의 token 값"
      }
    },
    "schemas": {
...
      "UserMe": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "point": { "type": "number" },
          "mileage": { "type": "number" },
          "notificationEnabled": { "type": "boolean" },
          "marketingConsent": { "type": "boolean" }
        }
      },
      "UserOrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "orderNo": { "type": "number" },
          "orderNumber": { "type": "string" },
          "status": { "type": "string" },
          "totalAmount": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "items": { "type": "array" }
        }
      },
...
    }
  }
}